+++ cache/UaiCacheInterface.php
@@ -1,0 +1,9 @@
+<?php
+
+interface UaiCacheInterface
+{
+  public static function get($key_name);
+  
+  public static function set($key_name, $data);
+}
+